+++ src/types/factories/IWETHZapper__factory.ts
@@ -34,9 +34,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint16",
+        internalType: "uint256",
         name: "referralCode",
-        type: "uint16",
+        type: "uint256",
       },
     ],
     name: "depositWithReferral",
